postgres/pyproject.toml
[project]
name = "postgres-crud-example"
version = "0.1.0"
description = "CRUD example with PostgreSQL using Xyra"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "xyra>=0.1.3",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.29.0",
]